requirements.txt
customtkinter
pillow
setuptools